requirements.txt
python-telegram-bot==20.7
requests==2.31.0
ffmpeg-python==0.2.0
flask==3.0.2 